pi/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.4
aiohttp-cors==0.8.1
aiosignal==1.4.0
annotated-types==0.7.0
attrs==26.1.0
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.6
click==8.3.1
colorful==0.5.8
cryptography==46.0.6
diskcache==5.6.3
distlib==0.4.0
filelock==3.25.2
frozenlist==1.8.0
google-api-core==2.30.1
google-auth==2.49.1
googleapis-common-protos==1.73.1
grpcio==1.80.0
idna==3.11
importlib-metadata==8.7.1
jinja2==3.1.6
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
llama-cpp-python==0.3.19
markupsafe==3.0.3
msgpack==1.1.2
multidict==6.7.1
numpy==2.4.4
opencensus==0.11.4
opencensus-context==0.1.3
opentelemetry-api==1.40.0
opentelemetry-exporter-prometheus==0.61b0
opentelemetry-proto==1.40.0
opentelemetry-sdk==1.40.0
opentelemetry-semantic-conventions==0.61b0
packaging==26.0
platformdirs==4.9.4
prometheus-client==0.24.1
propcache==0.4.1
proto-plus==1.27.2
protobuf==6.33.6
py-spy==0.4.1
pyasn1==0.6.3
pyasn1-modules==0.4.2
pycparser==3.0
pydantic==2.12.5
pydantic-core==2.41.5
python-discovery==1.2.1
pyyaml==6.0.3
ray==2.54.0
referencing==0.37.0
requests==2.33.1
rpds-py==0.30.0
six==1.17.0
smart-open==7.5.1
typing-extensions==4.15.0
typing-inspection==0.4.2
urllib3==2.6.3
virtualenv==21.2.0
wrapt==2.1.2
yarl==1.23.0
zipp==3.23.0
